[log] Перенёс json в отдельный файл
